+++ ext_tables.php
@@ -1,5 +1,7 @@
 <?php
 
+use Rubb1\Skins\Hooks\PageRenderer;
+
 defined('TYPO3') or die();
 
 /**
@@ -25,5 +27,5 @@ if (TYPO3 == 'BE') {
     ';
 
     // add backend css
-    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_pagerenderer.php']['render-preProcess'][] = \Rubb1\Skins\Hooks\PageRenderer::class . '->darkmode';
+    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_pagerenderer.php']['render-preProcess'][] = PageRenderer::class . '->darkmode';
 }