+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -7,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [11.0.0] - 2023-05-07
+
 ### Changed
+
 first release for version 12
 
 Breaking changes:
@@ -18,25 +21,34 @@ It's not final version yet.
 I have to do some more compatibility checks and tests.
 
 ## [10.5.3] - 2022-10-19
+
 ### Fixed
+
 - Fix access to possibly undefined array key (fixes #5)
 
 ## [10.5.2] - 2022-05-26
+
 ### Changed
+
 - Update styles for the mask extension module
 - Verify access to the user configuration to prevent possible errors
 - Change the flush cache icon styling
 
 ## [10.5.1] - 2022-03-18
+
 ### Changed
+
 - Update README.md
 
 ### Fixed
+
 - Fix version number
 - Fix css styles
 
 ## [10.5.0] - 2022-03-18
+
 ### Added
+
 - Create .gitignore
 - Add localization support
 - Add new styles
@@ -44,17 +56,24 @@ I have to do some more compatibility checks and tests.
 - Create CHANGELOG.md
 
 ### Changed
+
 - Outsource fluid templates
 - Update README.md
 
 ### Fixed
+
 - Add missing typo3/cms-core version for typo3 11
 
 ## [10.4.28] - 2021-10-07
 
 [Unreleased]: https://github.com/Dominik-Sch/skins/compare/v10.5.3...HEAD
+
 [10.5.3]: https://github.com/Dominik-Sch/skins/compare/v10.5.2...10.5.3
+
 [10.5.2]: https://github.com/Dominik-Sch/skins/compare/v10.5.1...10.5.2
+
 [10.5.1]: https://github.com/Dominik-Sch/skins/compare/v10.5.0...10.5.1
+
 [10.5.0]: https://github.com/Dominik-Sch/skins/compare/v10.4.28...10.5.0
+
 [10.4.28]: https://github.com/Dominik-Sch/skins/releases/tag/v10.4.28